Added Dark Souls like Hot Keys.
Add hotkeys. In early stages ATM dragging a consumable on the top button bar or into the first one of the four slots on the bottom of the equip window will enable you to use the E key, or square button on a controller to use item.
Hitting H (no controller set up yet) will switch through the consumables slotted into the four slots on the bottom of the equip window.