Hotkey bar done.
HotkeyBar is complete. This update refined a lot and add a spell/abilities hotkey with attunement slots just like in Darks souls. Two spell/abilities can be bought at the merchant and equipped at the status statue.
I also finalized the controls for the hotkey bar, that are almost Identical to eldinring's for both keyboard and controller. (note to use ability/spell its mapped to mouse button 4 and GP L2+L1)