Made the first round of build UI
Added a Bar on the right side that show the keys and Icons bound to build mode.
I tried to make it so it give almost all the info you need but at the same time looks good without crowding the screen.
As such it is minimalistic.
Any thoughts on the UI would be great.

Now I am going to begin work on a selection wheel to filter and select build items.