Fixed known Bugs
I have fix all know bugs except for one lol
The only bug left that I know of is the one where if you fill up your inventory till it adds new slots then save and reload it will not show the added slots until you sort the inventory.

This will be the last update for a while because I am going on Vacation. When I get back I plan on finishing the building UI.