Mini Map
Added a Mini Map. This is a prelude to a actual map which will be need in my quest and Dialog system.
I also intent to use unreal's new built in conversation system which was interduce in a later version then 5.2.1 which this systems is built on.
So I will have to update to the latest UE version soon.