Added the basics for the dialogue system
You can now talk to the merchant. Its kind of broke right now because you can actually buy anything at the moment.
Also if the conversation does not proceed after he says he is also a merchant. Just open the master conversation graph and change the stage on the one on the right to 1. For some reason it is not saving the changes some times.