Reworked equipment coding
Before when equipping an item the variable controlling what was equipped was only denoted by class. This meant that it was impossible to equip two items of the same class "IE: a crusted blade in the left and right hand or have a leather bangle in more then one of your accessory slots"

Now the equipment component use a struct that has the class and a GUID so you can equip as many of the same class as you have equipment slots for.

I also added a few items and fixed a few bugs