Fix Quest Bugs
For some reason the check if equip type Interface function will not save no mater what I did.
I made it and it worked perfectly save and exit and next time I reloaded the project the function was empty. I tried 5 times! Some of which deleting it and naming it something ells and every time I exit the project and reload it, it was empty!!!
So I had to hard reference code it which has no true detrimental effect because if a conversation is active the quest log is most likely to be as well but it just aggravating to have to do it that way because of a UE bug.